% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volatility_metrics.R
\name{compute_all_metrics}
\alias{compute_all_metrics}
\title{Compute all metrics}
\usage{
compute_all_metrics(y, window_size = 20, overlapping = TRUE)
}
\arguments{
\item{y}{Time series (vector)}

\item{window_size}{scalar. Width of the window over which variances are calculated}

\item{overlapping}{boolean. Whether to use overlapping (Default = TRUE) or not overlapping chunks (FALSE)}
}
\description{
Wrapper function to calculate all metrics at once
}
