% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurewise_volatility.R
\name{get_fw_volatility}
\alias{get_fw_volatility}
\title{Make a table of pairwise feature differences based on group membership.}
\usage{
get_fw_volatility(pw_dist, f_table)
}
\arguments{
\item{pw_dist}{A pairwise distance table, for instance generated by \code{get_pairwise_distance}.}

\item{f_table}{A table containing the features (as columns) that should be contrasted over time.}
}
\value{
a \code{data.frame} with pairwise differences
}
\description{
Given the relevant pairwise distances, compute change in every feature in a table.  .
}
