% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_local_variances.R
\name{get_local_variances}
\alias{get_local_variances}
\title{Calculate local variances of a time series}
\usage{
get_local_variances(y, window_size = 20, overlapping = TRUE)
}
\arguments{
\item{y}{Time series (vector)}

\item{window_size}{scalar. Width of the window over which variances are calculated}

\item{overlapping}{boolean. Whether to use overlapping (Default = TRUE) or not overlapping chunks (FALSE)}
}
\value{
A vector of local variances
}
\description{
Estimates over either non overlapping chunks of the time series
or overlapping are available. Choice is often based on available time series length
}
