% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volatility_metrics.R
\name{metric_cv_var}
\alias{metric_cv_var}
\title{Coefficient of Variation of Local Variances (CV)Volatility Metrics}
\usage{
metric_cv_var(y, window_size = 20, overlapping = TRUE)
}
\arguments{
\item{y}{Time series (vector)}

\item{window_size}{scalar. Width of the window over which variances are calculated}

\item{overlapping}{boolean. Whether to use overlapping (Default = TRUE) or not overlapping chunks (FALSE)}
}
\value{
CV metric
}
\description{
Volatility metric: Calculates the CV of local variances
}
