% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volatility_metrics.R
\name{metric_iqr_log_var}
\alias{metric_iqr_log_var}
\title{Interquartile Range of Log-Variances (IQR_log)}
\usage{
metric_iqr_log_var(y, window_size = 20, overlapping = TRUE)
}
\arguments{
\item{y}{Time series (vector)}

\item{window_size}{scalar. Width of the window over which variances are calculated}

\item{overlapping}{boolean. Whether to use overlapping (Default = TRUE) or not overlapping chunks (FALSE)}
}
\value{
IQR metric
}
\description{
Volatility metric: Calculates the Interquartile range of the local variances
on the log scale
}
