% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulations.R
\name{run_simulations}
\alias{run_simulations}
\title{Perform Monte Carlo simulations}
\usage{
run_simulations(
  n,
  nsims,
  window_size,
  sigma_vec,
  lambda = FALSE,
  homo = TRUE,
  trend = c("none", "linear", "sinusoid", "both"),
  detrend = c("none", "linear", "diff", "rolling")
)
}
\description{
Simulation study
}
