% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{simulate time series.}
\usage{
simulate_data(n, mu = 0, sigma = 1, mu_t = FALSE, sigma_t = FALSE)
}
\arguments{
\item{n}{length of the time series}

\item{mu}{Mean value of the Gaussian error (Default = 0)}

\item{sigma}{Standar deviation of the Gaussian error (Default = 1)}

\item{mu_t}{Boolean. Should the time series incorporate a linear trend (Default = FALSE)}

\item{sigma_t}{Boolean Should the time series incorporate herteroskedasticity (Default = FALSE)}
}
\value{
vector
}
\description{
Either stationary time series, or non stationary (1st or second order)
non stationarity is coded to increase linearly with time
}
\section{Model}{

}

