% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vola_genus_table}
\alias{vola_genus_table}
\title{The untransformed genus-level count table for the volatility study}
\format{
A data.frame object with 137 rows, genera, and 120 columns, samples.
}
\source{
\url{https://doi.org/10.1016/j.psyneuen.2020.105047}
}
\usage{
vola_genus_table
}
\description{
DADA2 with SILVA138 was used to generate this count table.
}
\keyword{datasets}
