% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volatility.R
\name{volatility}
\alias{volatility}
\title{Compute volatility between pairs of microbiome count data}
\usage{
volatility(counts, metadata, transform = TRUE, verbose = TRUE)
}
\arguments{
\item{counts}{a microbiome count table, with rows as features and columns as samples}

\item{metadata}{a vector in the same order as \code{counts}, containing information on which samples are from the same source and should be linked.}

\item{transform}{a boolean, whether to CLR-transform count data beforehand. Highly recommended.}

\item{verbose}{A boolean. Toggles whether to print diagnostic information while running. Useful for debugging errors on large datasets.}
}
\value{
a data.frame with IDs and volatility estimates per sample pair
}
\description{
The main function of the volatility package.
}
