% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clr_lite.R
\name{volatility_compute_clr}
\alias{volatility_compute_clr}
\title{compute CLR using Aitchison's method}
\usage{
volatility_compute_clr(x)
}
\arguments{
\item{x}{A vector of compositional data without zeroes.}
}
\value{
A vector of CLR-transformed data
}
\description{
See \code{clr_lite}.
}
\seealso{
\code{\link{clr_lite}}
}
