% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{simulate time series.}
\usage{
simulate_data(
  n,
  sigma = 1,
  lambda = FALSE,
  homo = TRUE,
  trend = c("none", "linear", "sinusoid", "both")
)
}
\arguments{
\item{n}{length of the time series}

\item{sigma}{Standar deviation of the Gaussian error (Default = 1)}

\item{lambda}{Boolean. Should the time series exponential growth in variance (Default = FALSE)}

\item{homo}{Boolean. Should the time series incorporate homoskedasticity (Default = TRUE)}

\item{trend}{Character. Form of trend: "none", "linear", "sinusoid", "both"}
}
\value{
vector
}
\description{
Either stationary time series, or non stationary (1st or second order)
non stationarity is coded to increase linearly with time
}
\section{Model}{

}

